Refactor SalesScreen with modular components and services
- Extract product and cart logic into dedicated services (ProductService, CartService)
- Create modular sales components: ProductGrid, ShoppingCart, PaymentDialog
- Simplify SalesScreen by delegating responsibilities to new components
- Improve code organization and separation of concerns
- Maintain existing functionality with new component structure
