+++ src/components/sales/index.js
@@ -1,0 +1,3 @@
+export { default as ProductGrid } from './ProductGrid';
+export { default as ShoppingCart } from './ShoppingCart';
+export { default as PaymentDialog } from './PaymentDialog';
