+++ src/pages/SalesScreen.jsx
@@ -1,40 +1,13 @@
 import React, { useState, useCallback, useMemo } from 'react';
-import { Box, Grid, Paper, Typography, Divider, Stack } from '../components/ui';
-import {
-  Button,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  TextField,
-  FormControl,
-  FormControlLabel,
-  RadioGroup,
-  Radio,
-  InputAdornment,
-  Slide,
-  Chip,
-  Snackbar,
-  Alert,
-} from '@mui/material';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
-import AddIcon from '@mui/icons-material/Add';
-import RemoveIcon from '@mui/icons-material/Remove';
-import DeleteIcon from '@mui/icons-material/Delete';
-import PaymentIcon from '@mui/icons-material/Payment';
-import ReceiptIcon from '@mui/icons-material/Receipt';
-import CreditCardIcon from '@mui/icons-material/CreditCard';
-import EuroIcon from '@mui/icons-material/Euro';
-import AccountBalanceIcon from '@mui/icons-material/AccountBalance';
-import LocalAtmIcon from '@mui/icons-material/LocalAtm';
-import CheckCircleIcon from '@mui/icons-material/CheckCircle';
-import CardGiftcardIcon from '@mui/icons-material/CardGiftcard';
-import SettingsIcon from '@mui/icons-material/Settings';
+import { Box, Grid, Typography, Alert } from '../components/ui';
+import { Slide, Snackbar } from '@mui/material';
 import {
   RedeemVoucherDialog,
   VoucherManagementDialog,
   PurchaseVoucherDialog,
 } from '../components/vouchers';
+import { ProductGrid, ShoppingCart, PaymentDialog } from '../components/sales';
+import { ProductService, CartService } from '../services';
 
 // Transition for dialog
 const Transition = React.forwardRef(function Transition(props, ref) {
@@ -46,24 +19,11 @@ const Transition = React.forwardRef(function Transition(props, ref) {
  * Displays products, cart, and payment options
  */
 const SalesScreen = () => {
-  // Sample product data - in a real app, this would come from an API
-  const products = useMemo(
-    () => [
-      { id: 1, name: 'Kaffee', price: 2.5, category: 'Getränke' },
-      { id: 2, name: 'Espresso', price: 1.8, category: 'Getränke' },
-      { id: 3, name: 'Cappuccino', price: 3.2, category: 'Getränke' },
-      { id: 4, name: 'Latte Macchiato', price: 3.5, category: 'Getränke' },
-      { id: 5, name: 'Tee', price: 2.0, category: 'Getränke' },
-      { id: 6, name: 'Croissant', price: 1.5, category: 'Gebäck' },
-      { id: 7, name: 'Muffin', price: 2.0, category: 'Gebäck' },
-      { id: 8, name: 'Bagel', price: 2.5, category: 'Gebäck' },
-      { id: 9, name: 'Sandwich', price: 4.0, category: 'Snacks' },
-      { id: 10, name: 'Salat', price: 5.5, category: 'Snacks' },
-      { id: 11, name: 'Suppe', price: 4.5, category: 'Snacks' },
-      { id: 12, name: 'Wasser', price: 1.0, category: 'Getränke' },
-    ],
-    []
-  );
+  // Get products from service
+  const products = useMemo(() => ProductService.getProducts(), []);
+
+  // Group products by category
+  const productsByCategory = useMemo(() => ProductService.groupByCategory(products), [products]);
 
   // State for cart items
   const [cartItems, setCartItems] = useState([]);
@@ -87,35 +47,17 @@ const SalesScreen = () => {
 
   // Add item to cart
   const addToCart = useCallback(product => {
-    setCartItems(prevItems => {
-      const existingItem = prevItems.find(item => item.id === product.id);
-      if (existingItem) {
-        return prevItems.map(item =>
-          item.id === product.id ? { ...item, quantity: item.quantity + 1 } : item
-        );
-      } else {
-        return [...prevItems, { ...product, quantity: 1 }];
-      }
-    });
+    setCartItems(prevItems => CartService.addToCart(prevItems, product));
   }, []);
 
   // Remove item from cart
   const removeFromCart = useCallback(productId => {
-    setCartItems(prevItems => {
-      const existingItem = prevItems.find(item => item.id === productId);
-      if (existingItem && existingItem.quantity > 1) {
-        return prevItems.map(item =>
-          item.id === productId ? { ...item, quantity: item.quantity - 1 } : item
-        );
-      } else {
-        return prevItems.filter(item => item.id !== productId);
-      }
-    });
+    setCartItems(prevItems => CartService.removeFromCart(prevItems, productId));
   }, []);
 
   // Delete item from cart
   const deleteFromCart = useCallback(productId => {
-    setCartItems(prevItems => prevItems.filter(item => item.id !== productId));
+    setCartItems(prevItems => CartService.deleteFromCart(prevItems, productId));
   }, []);
 
   // Clear cart
@@ -125,31 +67,19 @@ const SalesScreen = () => {
   }, []);
 
   // Calculate subtotal (before vouchers)
-  const subtotal = useMemo(() => {
-    return cartItems.reduce((sum, item) => sum + item.price * item.quantity, 0);
-  }, [cartItems]);
+  const subtotal = useMemo(() => CartService.calculateSubtotal(cartItems), [cartItems]);
 
   // Calculate voucher discount
-  const voucherDiscount = useMemo(() => {
-    return appliedVouchers.reduce((sum, voucher) => sum + voucher.value, 0);
-  }, [appliedVouchers]);
+  const voucherDiscount = useMemo(
+    () => CartService.calculateVoucherDiscount(appliedVouchers),
+    [appliedVouchers]
+  );
 
   // Calculate total (after vouchers)
-  const total = useMemo(() => {
-    return Math.max(0, subtotal - voucherDiscount);
-  }, [subtotal, voucherDiscount]);
-
-  // Group products by category
-  const productsByCategory = useMemo(() => {
-    const grouped = {};
-    products.forEach(product => {
-      if (!grouped[product.category]) {
-        grouped[product.category] = [];
-      }
-      grouped[product.category].push(product);
-    });
-    return grouped;
-  }, [products]);
+  const total = useMemo(
+    () => CartService.calculateTotal(subtotal, voucherDiscount),
+    [subtotal, voucherDiscount]
+  );
 
   // Handle payment modal open
   const handlePaymentModalOpen = useCallback(() => {
@@ -253,501 +183,48 @@ const SalesScreen = () => {
       <Grid container spacing={3} sx={{ height: 'calc(100% - 60px)' }}>
         {/* Product selection area */}
         <Grid item xs={12} md={8}>
-          <Paper
-            elevation={3}
-            sx={{
-              p: 2,
-              height: '100%',
-              overflow: 'auto',
-              bgcolor: 'background.paper',
-            }}
-          >
-            <Typography variant="h5" gutterBottom>
-              Produkte
-            </Typography>
-
-            {Object.entries(productsByCategory).map(([category, categoryProducts]) => (
-              <Box key={category} sx={{ mb: 3 }}>
-                <Typography
-                  variant="h6"
-                  sx={{
-                    mb: 1,
-                    pb: 0.5,
-                    borderBottom: '1px solid',
-                    borderColor: 'divider',
-                  }}
-                >
-                  {category}
-                </Typography>
-
-                <Grid container spacing={2}>
-                  {categoryProducts.map(product => (
-                    <Grid item xs={6} sm={4} md={3} key={product.id}>
-                      <Paper
-                        elevation={2}
-                        sx={{
-                          p: 2,
-                          height: '100%',
-                          display: 'flex',
-                          flexDirection: 'column',
-                          justifyContent: 'space-between',
-                          transition: 'all 0.2s',
-                          '&:hover': {
-                            transform: 'translateY(-4px)',
-                            boxShadow: 4,
-                          },
-                          cursor: 'pointer',
-                        }}
-                        onClick={() => addToCart(product)}
-                      >
-                        <Typography variant="subtitle1" component="div" fontWeight="medium">
-                          {product.name}
-                        </Typography>
-                        <Typography variant="body2" color="text.secondary">
-                          {product.price.toFixed(2)} €
-                        </Typography>
-                      </Paper>
-                    </Grid>
-                  ))}
-                </Grid>
-              </Box>
-            ))}
-          </Paper>
+          <ProductGrid productsByCategory={productsByCategory} onProductSelect={addToCart} />
         </Grid>
 
         {/* Cart area */}
         <Grid item xs={12} md={4}>
-          <Paper
-            elevation={3}
-            sx={{
-              p: 2,
-              height: '100%',
-              display: 'flex',
-              flexDirection: 'column',
-              bgcolor: 'background.paper',
-            }}
-          >
-            <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-              <ShoppingCartIcon sx={{ mr: 1 }} />
-              <Typography variant="h5">Warenkorb</Typography>
-              {cartItems.length > 0 && (
-                <Chip
-                  label={cartItems.reduce((sum, item) => sum + item.quantity, 0)}
-                  color="primary"
-                  size="small"
-                  sx={{ ml: 1 }}
-                />
-              )}
-            </Box>
-
-            <Divider sx={{ mb: 2 }} />
-
-            {/* Cart items */}
-            <Box sx={{ flexGrow: 1, overflow: 'auto', mb: 2 }}>
-              {cartItems.length === 0 ? (
-                <Typography
-                  variant="body1"
-                  color="text.secondary"
-                  sx={{ textAlign: 'center', mt: 4 }}
-                >
-                  Der Warenkorb ist leer
-                </Typography>
-              ) : (
-                <Stack spacing={2}>
-                  {cartItems.map(item => (
-                    <Paper key={item.id} variant="outlined" sx={{ p: 1.5 }}>
-                      <Box sx={{ display: 'flex', justifyContent: 'space-between', mb: 1 }}>
-                        <Typography variant="subtitle1">{item.name}</Typography>
-                        <Typography variant="subtitle1">
-                          {(item.price * item.quantity).toFixed(2)} €
-                        </Typography>
-                      </Box>
-
-                      <Box
-                        sx={{
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'space-between',
-                        }}
-                      >
-                        <Typography variant="body2" color="text.secondary">
-                          {item.price.toFixed(2)} € × {item.quantity}
-                        </Typography>
-
-                        <Box>
-                          <Button
-                            size="small"
-                            onClick={() => removeFromCart(item.id)}
-                            sx={{ minWidth: '32px', p: 0.5 }}
-                          >
-                            <RemoveIcon fontSize="small" />
-                          </Button>
-
-                          <Button
-                            size="small"
-                            onClick={() => addToCart(item)}
-                            sx={{ minWidth: '32px', p: 0.5 }}
-                          >
-                            <AddIcon fontSize="small" />
-                          </Button>
-
-                          <Button
-                            size="small"
-                            color="error"
-                            onClick={() => deleteFromCart(item.id)}
-                            sx={{ minWidth: '32px', p: 0.5 }}
-                          >
-                            <DeleteIcon fontSize="small" />
-                          </Button>
-                        </Box>
-                      </Box>
-                    </Paper>
-                  ))}
-
-                  {/* Applied vouchers */}
-                  {appliedVouchers.length > 0 && (
-                    <Box sx={{ mt: 2 }}>
-                      <Typography variant="subtitle2" gutterBottom>
-                        Angewendete Gutscheine:
-                      </Typography>
-                      <Stack spacing={1}>
-                        {appliedVouchers.map(voucher => (
-                          <Paper
-                            key={voucher.id}
-                            variant="outlined"
-                            sx={{ p: 1.5, bgcolor: 'background.default' }}
-                          >
-                            <Box sx={{ display: 'flex', justifyContent: 'space-between', mb: 0.5 }}>
-                              <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                                <CardGiftcardIcon
-                                  fontSize="small"
-                                  sx={{ mr: 1, color: 'primary.main' }}
-                                />
-                                <Typography variant="body2">{voucher.code}</Typography>
-                              </Box>
-                              <Typography variant="body2" color="error.main" fontWeight="medium">
-                                -{voucher.value.toFixed(2)} €
-                              </Typography>
-                            </Box>
-                            <Box sx={{ display: 'flex', justifyContent: 'flex-end' }}>
-                              <Button
-                                size="small"
-                                color="error"
-                                onClick={() => handleRemoveVoucher(voucher.id)}
-                                sx={{ minWidth: '32px', p: 0.5 }}
-                              >
-                                <DeleteIcon fontSize="small" />
-                              </Button>
-                            </Box>
-                          </Paper>
-                        ))}
-                      </Stack>
-                    </Box>
-                  )}
-                </Stack>
-              )}
-            </Box>
-
-            {/* Cart summary */}
-            <Box>
-              <Divider sx={{ mb: 2 }} />
-              <Button
-                variant="outlined"
-                color="primary"
-                startIcon={<CardGiftcardIcon />}
-                onClick={handlePurchaseVoucherDialogOpen}
-                size="small"
-                fullWidth
-              >
-                Gutschein kaufen
-              </Button>
-
-              {/* Voucher buttons */}
-              {!receiptReady && cartItems.length > 0 && (
-                <Box sx={{ display: 'flex', gap: 1, mb: 2 }}>
-                  <Button
-                    variant="outlined"
-                    color="primary"
-                    startIcon={<CardGiftcardIcon />}
-                    onClick={handleRedeemVoucherDialogOpen}
-                    size="small"
-                    fullWidth
-                  >
-                    Gutschein einlösen
-                  </Button>
-                  <Button
-                    variant="outlined"
-                    color="primary"
-                    startIcon={<SettingsIcon />}
-                    onClick={handleVoucherManagementDialogOpen}
-                    size="small"
-                  >
-                    Verwalten
-                  </Button>
-                </Box>
-              )}
-
-              {/* Subtotal and discount */}
-              {voucherDiscount > 0 && (
-                <>
-                  <Box sx={{ display: 'flex', justifyContent: 'space-between', mb: 1 }}>
-                    <Typography variant="body1">Zwischensumme:</Typography>
-                    <Typography variant="body1">{subtotal.toFixed(2)} €</Typography>
-                  </Box>
-                  <Box sx={{ display: 'flex', justifyContent: 'space-between', mb: 1 }}>
-                    <Typography variant="body1" color="error.main">
-                      Gutschein-Rabatt:
-                    </Typography>
-                    <Typography variant="body1" color="error.main">
-                      -{voucherDiscount.toFixed(2)} €
-                    </Typography>
-                  </Box>
-                </>
-              )}
-
-              {/* Total */}
-              <Box sx={{ display: 'flex', justifyContent: 'space-between', mb: 2 }}>
-                <Typography variant="h6">Gesamtsumme:</Typography>
-                <Typography variant="h6">{total.toFixed(2)} €</Typography>
-              </Box>
-
-              {receiptReady ? (
-                <Stack spacing={2}>
-                  <Button
-                    variant="contained"
-                    color="success"
-                    fullWidth
-                    startIcon={<ReceiptIcon />}
-                    onClick={handlePrintReceipt}
-                  >
-                    Rechnung drucken
-                  </Button>
-
-                  <Button
-                    variant="outlined"
-                    color="primary"
-                    fullWidth
-                    startIcon={<ShoppingCartIcon />}
-                    onClick={handleNewTransaction}
-                  >
-                    Neuer Verkauf
-                  </Button>
-                </Stack>
-              ) : (
-                <Stack direction="row" spacing={2}>
-                  <Button
-                    variant="outlined"
-                    color="error"
-                    fullWidth
-                    startIcon={<DeleteIcon />}
-                    onClick={clearCart}
-                    disabled={cartItems.length === 0}
-                  >
-                    Leeren
-                  </Button>
-
-                  <Button
-                    variant="contained"
-                    color="primary"
-                    fullWidth
-                    startIcon={<PaymentIcon />}
-                    onClick={handlePaymentModalOpen}
-                    disabled={cartItems.length === 0}
-                  >
-                    Bezahlen
-                  </Button>
-                </Stack>
-              )}
-            </Box>
-          </Paper>
+          <ShoppingCart
+            cartItems={cartItems}
+            appliedVouchers={appliedVouchers}
+            subtotal={subtotal}
+            voucherDiscount={voucherDiscount}
+            total={total}
+            receiptReady={receiptReady}
+            onAddItem={addToCart}
+            onRemoveItem={removeFromCart}
+            onDeleteItem={deleteFromCart}
+            onClearCart={clearCart}
+            onPayment={handlePaymentModalOpen}
+            onPrintReceipt={handlePrintReceipt}
+            onNewTransaction={handleNewTransaction}
+            onRemoveVoucher={handleRemoveVoucher}
+            onRedeemVoucher={handleRedeemVoucherDialogOpen}
+            onManageVouchers={handleVoucherManagementDialogOpen}
+            onPurchaseVoucher={handlePurchaseVoucherDialogOpen}
+          />
         </Grid>
       </Grid>
 
-      {/* Payment Modal */}
-      <Dialog
+      {/* Payment Dialog */}
+      <PaymentDialog
         open={paymentModalOpen}
         onClose={handlePaymentModalClose}
+        onComplete={handlePaymentComplete}
+        total={total}
+        cartItemsCount={cartItems.length}
+        voucherDiscount={voucherDiscount}
+        paymentMethod={paymentMethod}
+        onPaymentMethodChange={handlePaymentMethodChange}
+        cashReceived={cashReceived}
+        onCashReceivedChange={handleCashReceivedChange}
+        change={change}
         TransitionComponent={Transition}
-        maxWidth="sm"
-        fullWidth
-      >
-        <DialogTitle>
-          <Box sx={{ display: 'flex', alignItems: 'center' }}>
-            <PaymentIcon sx={{ mr: 1 }} />
-            Zahlung abschließen
-          </Box>
-        </DialogTitle>
-
-        <DialogContent dividers>
-          <Box sx={{ mb: 3 }}>
-            <Typography variant="h6" gutterBottom>
-              Gesamtbetrag: {total.toFixed(2)} €
-            </Typography>
-
-            <Typography variant="body2" color="text.secondary">
-              {cartItems.length} {cartItems.length === 1 ? 'Artikel' : 'Artikel'} im Warenkorb
-            </Typography>
-
-            {voucherDiscount > 0 && (
-              <Typography variant="body2" color="error.main" sx={{ mt: 1 }}>
-                Gutschein-Rabatt: {voucherDiscount.toFixed(2)} €
-              </Typography>
-            )}
-          </Box>
-
-          <FormControl component="fieldset" sx={{ mb: 3 }}>
-            <Typography variant="subtitle1" gutterBottom>
-              Zahlungsmethode
-            </Typography>
-
-            <RadioGroup
-              name="payment-method"
-              value={paymentMethod}
-              onChange={handlePaymentMethodChange}
-            >
-              <Paper variant="outlined" sx={{ mb: 1, p: 1 }}>
-                <FormControlLabel
-                  value="cash"
-                  control={<Radio />}
-                  label={
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                      <LocalAtmIcon sx={{ mr: 1 }} />
-                      <Typography>Bargeld</Typography>
-                    </Box>
-                  }
-                />
-              </Paper>
-
-              <Paper variant="outlined" sx={{ mb: 1, p: 1 }}>
-                <FormControlLabel
-                  value="card"
-                  control={<Radio />}
-                  label={
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                      <CreditCardIcon sx={{ mr: 1 }} />
-                      <Typography>Kartenzahlung</Typography>
-                    </Box>
-                  }
-                />
-              </Paper>
-
-              <Paper variant="outlined" sx={{ p: 1 }}>
-                <FormControlLabel
-                  value="bank"
-                  control={<Radio />}
-                  label={
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                      <AccountBalanceIcon sx={{ mr: 1 }} />
-                      <Typography>Überweisung</Typography>
-                    </Box>
-                  }
-                />
-              </Paper>
-            </RadioGroup>
-          </FormControl>
-
-          {paymentMethod === 'cash' && (
-            <Box sx={{ mb: 2 }}>
-              <Typography variant="subtitle1" gutterBottom>
-                Bargeld
-              </Typography>
-
-              <TextField
-                label="Erhaltener Betrag"
-                variant="outlined"
-                fullWidth
-                value={cashReceived}
-                onChange={handleCashReceivedChange}
-                type="number"
-                InputProps={{
-                  startAdornment: (
-                    <InputAdornment position="start">
-                      <EuroIcon />
-                    </InputAdornment>
-                  ),
-                }}
-                sx={{ mb: 2 }}
-              />
-
-              <Paper
-                variant="outlined"
-                sx={{
-                  p: 2,
-                  bgcolor: 'background.default',
-                  display: 'flex',
-                  justifyContent: 'space-between',
-                  alignItems: 'center',
-                }}
-              >
-                <Typography variant="subtitle1">Rückgeld:</Typography>
-                <Typography variant="h6" color="primary.main">
-                  {change} €
-                </Typography>
-              </Paper>
-            </Box>
-          )}
-
-          {paymentMethod === 'card' && (
-            <Box sx={{ mb: 2 }}>
-              <Typography variant="body1" sx={{ mb: 2 }}>
-                Bitte führen Sie die Karte in das Lesegerät ein oder halten Sie sie an das Terminal.
-              </Typography>
-
-              <Box
-                sx={{
-                  display: 'flex',
-                  justifyContent: 'center',
-                  alignItems: 'center',
-                  p: 3,
-                  border: '1px dashed',
-                  borderColor: 'divider',
-                  borderRadius: 1,
-                }}
-              >
-                <CreditCardIcon sx={{ fontSize: 60, color: 'text.secondary', opacity: 0.7 }} />
-              </Box>
-            </Box>
-          )}
-
-          {paymentMethod === 'bank' && (
-            <Box sx={{ mb: 2 }}>
-              <Typography variant="body1" gutterBottom>
-                Bitte überweisen Sie den Betrag auf folgendes Konto:
-              </Typography>
-
-              <Paper variant="outlined" sx={{ p: 2, bgcolor: 'background.default' }}>
-                <Typography variant="body2" gutterBottom>
-                  <strong>Kontoinhaber:</strong> Vendura GmbH
-                </Typography>
-                <Typography variant="body2" gutterBottom>
-                  <strong>IBAN:</strong> DE12 3456 7890 1234 5678 90
-                </Typography>
-                <Typography variant="body2" gutterBottom>
-                  <strong>BIC:</strong> DEUTDEDBXXX
-                </Typography>
-                <Typography variant="body2">
-                  <strong>Verwendungszweck:</strong> Rechnung{' '}
-                  {new Date().toISOString().slice(0, 10)}-{Math.floor(Math.random() * 1000)}
-                </Typography>
-              </Paper>
-            </Box>
-          )}
-        </DialogContent>
-
-        <DialogActions sx={{ p: 2 }}>
-          <Button onClick={handlePaymentModalClose} color="inherit">
-            Abbrechen
-          </Button>
-          <Button
-            onClick={handlePaymentComplete}
-            variant="contained"
-            color="primary"
-            startIcon={<CheckCircleIcon />}
-          >
-            Zahlung abschließen
-          </Button>
-        </DialogActions>
-      </Dialog>
+      />
 
       {/* Success Snackbar */}
       <Snackbar
