+++ src/services/index.js
@@ -1,0 +1,2 @@
+export { default as ProductService } from './ProductService';
+export { default as CartService } from './CartService';
